fix circular graph
