+++ packages/kbn-generate-csv/tsconfig.json
@@ -24,7 +24,6 @@
     "@kbn/config-schema",
     "@kbn/i18n",
     "@kbn/reporting-common",
-    "@kbn/reporting-export-types-csv-common",
     "@kbn/reporting-server",
   ]
 }
