+++ packages/kbn-generate-csv/types.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import type { SerializedSearchSourceFields } from '@kbn/data-plugin/public';
+
+/**
+ * Duplicated from @kbn/reporting-export-types-csv-common to reduce dependencies
+ */
+export interface JobParamsCSV {
+  browserTimezone?: string;
+  searchSource: SerializedSearchSourceFields;
+  columns?: string[];
+}
