+++ packages/kbn-reporting/common/constants.ts
@@ -8,27 +8,6 @@
 
 export const PLUGIN_ID = 'reporting';
 
-// Export Type Sets
-import {
-  CSV_JOB_TYPE,
-  CSV_JOB_TYPE_DEPRECATED,
-  CSV_JOB_TYPE_V2,
-  CSV_REPORT_TYPE,
-  CSV_REPORT_TYPE_V2,
-} from '@kbn/reporting-export-types-csv-common';
-import {
-  PDF_JOB_TYPE,
-  PDF_JOB_TYPE_V2,
-  PDF_REPORT_TYPE,
-  PDF_REPORT_TYPE_V2,
-} from '@kbn/reporting-export-types-pdf-common';
-import {
-  PNG_JOB_TYPE,
-  PNG_JOB_TYPE_V2,
-  PNG_REPORT_TYPE,
-  PNG_REPORT_TYPE_V2,
-} from '@kbn/reporting-export-types-png-common';
-
 export const ALLOWED_JOB_CONTENT_TYPES = [
   'application/json',
   'application/pdf',
@@ -36,39 +15,6 @@ export const ALLOWED_JOB_CONTENT_TYPES = [
   'image/png',
   'text/plain',
 ];
-
-export const reportTypes = [
-  CSV_REPORT_TYPE,
-  CSV_REPORT_TYPE_V2,
-  PDF_REPORT_TYPE,
-  PDF_REPORT_TYPE_V2,
-  PNG_REPORT_TYPE,
-  PNG_REPORT_TYPE_V2,
-];
-
-export const jobTypes = [
-  CSV_JOB_TYPE,
-  CSV_JOB_TYPE_V2,
-  PDF_JOB_TYPE,
-  PDF_JOB_TYPE_V2,
-  PNG_JOB_TYPE,
-  PNG_JOB_TYPE_V2,
-];
-
-type ReportTypeDeclaration = typeof reportTypes;
-export type ReportTypes = ReportTypeDeclaration[keyof ReportTypeDeclaration];
-
-type JobTypeDeclaration = typeof jobTypes;
-export type JobTypes = JobTypeDeclaration[keyof JobTypeDeclaration];
-
-export const USES_HEADLESS_JOB_TYPES = [
-  PDF_JOB_TYPE,
-  PNG_JOB_TYPE,
-  PDF_JOB_TYPE_V2,
-  PNG_JOB_TYPE_V2,
-];
-
-export const DEPRECATED_JOB_TYPES = [CSV_JOB_TYPE_DEPRECATED];
 
 // APM
 export const REPORTING_TRANSACTION_TYPE = PLUGIN_ID;
