+++ packages/kbn-reporting/common/tsconfig.json
@@ -19,8 +19,5 @@
     "@kbn/utility-types",
     "@kbn/screenshotting-plugin",
     "@kbn/i18n",
-    "@kbn/reporting-export-types-csv-common",
-    "@kbn/reporting-export-types-pdf-common",
-    "@kbn/reporting-export-types-png-common",
   ]
 }
