+++ x-pack/plugins/reporting/public/lib/job.test.ts
@@ -5,8 +5,7 @@
  * 2.0.
  */
 
-import { jobTypes } from '@kbn/reporting-common';
-
+import { jobTypes } from '../../common/constants';
 import { mockJobs } from '../../common/test';
 import { Job } from './job';
