+++ x-pack/plugins/reporting/server/usage/get_export_stats.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { DEPRECATED_JOB_TYPES } from '@kbn/reporting-common';
+import { DEPRECATED_JOB_TYPES } from '../../common/constants';
 import { ExportTypesHandler } from './get_export_type_handler';
 import {
   AvailableTotal,
